Refactor: extract websocket connect logic
