+++ src/sinks/websocket/sink.rs
@@ -1,7 +1,5 @@
 use std::{
-    fmt::Debug,
     io,
-    net::SocketAddr,
     num::NonZeroU64,
     task::{Context, Poll},
     time::{Duration, Instant},
@@ -10,18 +8,10 @@ use std::{
 use async_trait::async_trait;
 use bytes::BytesMut;
 use futures::{pin_mut, sink::SinkExt, stream::BoxStream, Sink, Stream, StreamExt};
-use snafu::{ResultExt, Snafu};
-use tokio::{net::TcpStream, time};
-use tokio_tungstenite::{
-    client_async_with_config,
-    tungstenite::{
-        client::{uri_mode, IntoClientRequest},
-        error::{Error as WsError, ProtocolError, UrlError},
-        handshake::client::Request as WsRequest,
-        protocol::{Message, WebSocketConfig},
-        stream::Mode as UriMode,
-    },
-    WebSocketStream as WsStream,
+use tokio::time;
+use tokio_tungstenite::tungstenite::{
+    error::{Error as WsError, ProtocolError},
+    protocol::Message,
 };
 use tokio_util::codec::Encoder as _;
 use vector_core::{
@@ -33,139 +23,13 @@ use vector_core::{
 
 use crate::{
     codecs::{Encoder, Transformer},
-    dns, emit,
+    common::websocket::WebSocketConnector,
+    emit,
     event::{Event, EventStatus, Finalizable},
-    http::Auth,
-    internal_events::{
-        ConnectionOpen, OpenGauge, WsConnectionError, WsConnectionEstablished,
-        WsConnectionFailedError, WsConnectionShutdown,
-    },
-    sinks::util::{retries::ExponentialBackoff, StreamSink},
+    internal_events::{ConnectionOpen, OpenGauge, WsConnectionError, WsConnectionShutdown},
+    sinks::util::StreamSink,
     sinks::websocket::config::WebSocketSinkConfig,
-    tls::{MaybeTlsSettings, MaybeTlsStream, TlsError},
 };
-
-#[derive(Debug, Snafu)]
-#[snafu(visibility(pub))]
-pub enum WebSocketError {
-    #[snafu(display("Creating WebSocket client failed: {}", source))]
-    CreateFailed { source: WsError },
-    #[snafu(display("Connect error: {}", source))]
-    ConnectError { source: TlsError },
-    #[snafu(display("Unable to resolve DNS: {}", source))]
-    DnsError { source: dns::DnsError },
-    #[snafu(display("No addresses returned."))]
-    NoAddresses,
-}
-
-#[derive(Clone)]
-pub struct WebSocketConnector {
-    uri: String,
-    host: String,
-    port: u16,
-    tls: MaybeTlsSettings,
-    auth: Option<Auth>,
-}
-
-impl WebSocketConnector {
-    pub fn new(
-        uri: String,
-        tls: MaybeTlsSettings,
-        auth: Option<Auth>,
-    ) -> Result<Self, WebSocketError> {
-        let request = (&uri).into_client_request().context(CreateFailedSnafu)?;
-        let (host, port) = Self::extract_host_and_port(&request).context(CreateFailedSnafu)?;
-
-        Ok(Self {
-            uri,
-            host,
-            port,
-            tls,
-            auth,
-        })
-    }
-
-    fn extract_host_and_port(request: &WsRequest) -> Result<(String, u16), WsError> {
-        let host = request
-            .uri()
-            .host()
-            .ok_or(WsError::Url(UrlError::NoHostName))?
-            .to_string();
-        let mode = uri_mode(request.uri())?;
-        let port = request.uri().port_u16().unwrap_or(match mode {
-            UriMode::Tls => 443,
-            UriMode::Plain => 80,
-        });
-
-        Ok((host, port))
-    }
-
-    const fn fresh_backoff() -> ExponentialBackoff {
-        ExponentialBackoff::from_millis(2)
-            .factor(250)
-            .max_delay(Duration::from_secs(60))
-    }
-
-    async fn tls_connect(&self) -> Result<MaybeTlsStream<TcpStream>, WebSocketError> {
-        let ip = dns::Resolver
-            .lookup_ip(self.host.clone())
-            .await
-            .context(DnsSnafu)?
-            .next()
-            .ok_or(WebSocketError::NoAddresses)?;
-
-        let addr = SocketAddr::new(ip, self.port);
-        self.tls
-            .connect(&self.host, &addr)
-            .await
-            .context(ConnectSnafu)
-    }
-
-    async fn connect(&self) -> Result<WsStream<MaybeTlsStream<TcpStream>>, WebSocketError> {
-        let mut request = (&self.uri)
-            .into_client_request()
-            .context(CreateFailedSnafu)?;
-
-        if let Some(auth) = &self.auth {
-            auth.apply(&mut request);
-        }
-
-        let maybe_tls = self.tls_connect().await?;
-
-        let ws_config = WebSocketConfig {
-            max_send_queue: None, // don't buffer messages
-            ..Default::default()
-        };
-
-        let (ws_stream, _response) = client_async_with_config(request, maybe_tls, Some(ws_config))
-            .await
-            .context(CreateFailedSnafu)?;
-
-        Ok(ws_stream)
-    }
-
-    async fn connect_backoff(&self) -> WsStream<MaybeTlsStream<TcpStream>> {
-        let mut backoff = Self::fresh_backoff();
-        loop {
-            match self.connect().await {
-                Ok(ws_stream) => {
-                    emit!(WsConnectionEstablished {});
-                    return ws_stream;
-                }
-                Err(error) => {
-                    emit!(WsConnectionFailedError {
-                        error: Box::new(error)
-                    });
-                    time::sleep(backoff.next().unwrap()).await;
-                }
-            }
-        }
-    }
-
-    pub async fn healthcheck(&self) -> crate::Result<()> {
-        self.connect().await.map(|_| ()).map_err(Into::into)
-    }
-}
 
 struct PingInterval {
     interval: Option<time::Interval>,
@@ -199,7 +63,10 @@ pub struct WebSocketSink {
 }
 
 impl WebSocketSink {
-    pub fn new(config: &WebSocketSinkConfig, connector: WebSocketConnector) -> crate::Result<Self> {
+    pub(crate) fn new(
+        config: &WebSocketSinkConfig,
+        connector: WebSocketConnector,
+    ) -> crate::Result<Self> {
         let transformer = config.encoding.transformer();
         let serializer = config.encoding.build()?;
         let encoder = Encoder::<()>::new(serializer);
@@ -380,18 +247,18 @@ mod tests {
     use tokio::time::timeout;
     use tokio_tungstenite::{
         accept_async, accept_hdr_async,
-        tungstenite::error::{Error as WsError, ProtocolError},
         tungstenite::handshake::server::{Request, Response},
     };
 
     use super::*;
     use crate::{
         config::{SinkConfig, SinkContext},
+        http::Auth,
         test_util::{
             components::{run_and_assert_sink_compliance, SINK_TAGS},
             next_addr, random_lines_with_stream, trace_init, CountReceiver,
         },
-        tls::{self, TlsConfig, TlsEnableableConfig},
+        tls::{self, MaybeTlsSettings, TlsConfig, TlsEnableableConfig},
     };
 
     #[tokio::test(flavor = "multi_thread")]
